+++ R/read_envi.R
@@ -28,7 +28,7 @@
 #' @author Zacharias Steinmetz, Claudia Beleites
 #'
 #' @seealso
-#' \code{\link{read_spec}()} for reading .y(a)ml, .json, or .rds (OpenSpecy)
+#' \code{\link{read_spec}()} for reading .json, .rds, or .csv (OpenSpecy)
 #' files;
 #' \code{\link{read_text}()}, \code{\link{read_asp}()}, \code{\link{read_spa}()},
 #' \code{\link{read_spc}()}, and \code{\link{read_jdx}()} for text files, .asp,