+++ R/zzz.R
@@ -1,8 +1,8 @@
 #' @importFrom utils packageVersion
 if (getRversion() >= "2.15.1") {
   utils::globalVariables(c(
-    ".", ".N", ".rs.invokeShinyWindowExternal", "column", "correct", "level",
-    "n", "problem", "value"
+    ".", ".N", ".row_id", ".rs.invokeShinyWindowExternal", "column",
+    "correct", "level", "n", "problem", "value", "y"
   ))
 }